checkpoint: clippy, cleanup
